requirements.txt
asttokens==3.0.1
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
Faker==40.1.0
findspark==2.0.1
ipykernel==7.1.0
ipython==8.38.0
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
Pygments==2.19.2
pyspark==4.1.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.2.14
